feat: merge require and prefer node with cache into one file
