+++ pkg/webhook/plugins/plugins.go
@@ -20,9 +20,8 @@ import (
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/base"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/fusesidecar"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/mountpropagationinjector"
-	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/prefernodeswithcache"
+	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/nodeaffinitywithcache"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/prefernodeswithoutcache"
-	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/requirenodeswithcache"
 	"github.com/fluid-cloudnative/fluid/pkg/webhook/plugins/requirenodewithfuse"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -65,8 +64,7 @@ func Registry(client client.Client) plugins {
 		},
 		podWithDatasetHandler: []MutatingHandler{
 			requirenodewithfuse.NewPlugin(client),
-			requirenodeswithcache.NewPlugin(client),
-			prefernodeswithcache.NewPlugin(client),
+			nodeaffinitywithcache.NewPlugin(client),
 			mountpropagationinjector.NewPlugin(client),
 		}, serverlessPodHandler: []MutatingHandler{
 			fusesidecar.NewPlugin(client),
