--- pkg/webhook/plugins/requirenodeswithcache/require_nodes_with_cache_test.go
@@ -1,89 +1,0 @@
-/*
-Copyright 2021 The Fluid Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package requirenodeswithcache
-
-import (
-	"testing"
-
-	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
-	"github.com/fluid-cloudnative/fluid/pkg/ddc/base"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-func TestMutate(t *testing.T) {
-	var (
-		client     client.Client
-		noSchedPod *corev1.Pod
-		schedPod   *corev1.Pod
-	)
-
-	plugin := NewPlugin(client)
-	if plugin.GetName() != NAME {
-		t.Errorf("GetName expect %v, got %v", NAME, plugin.GetName())
-	}
-
-	runtimeInfo, err := base.BuildRuntimeInfo("test", "fluid", "alluxio", datav1alpha1.TieredStore{})
-	if err != nil {
-		t.Errorf("fail to create the runtimeInfo with error %v", err)
-	}
-
-	noSchedPod = &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-		},
-	}
-
-	shouldStop, err := plugin.Mutate(noSchedPod, map[string]base.RuntimeInfoInterface{"pvcName": runtimeInfo})
-	if err != nil {
-		t.Errorf("fail to mutate pod with error %v", err)
-	}
-
-	if shouldStop {
-		t.Errorf("expect shouldStop as false, but got %v", shouldStop)
-	}
-
-	_, err = plugin.Mutate(noSchedPod, map[string]base.RuntimeInfoInterface{})
-	if err != nil {
-		t.Errorf("fail to mutate pod with error %v", err)
-	}
-
-	schedPod = &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-			Labels: map[string]string{
-				"fluid.io/dataset.test10-ds.sched": "required",
-			},
-		},
-	}
-
-	_, err = plugin.Mutate(schedPod, map[string]base.RuntimeInfoInterface{"pvcName": nil})
-	if err != nil || schedPod.Spec.Affinity != nil {
-		t.Errorf("expect error is nil and no injection")
-	}
-
-	_, err = plugin.Mutate(schedPod, map[string]base.RuntimeInfoInterface{"test10-ds": runtimeInfo})
-	if err != nil {
-		t.Errorf("fail to mutate pod with error %v", err)
-	}
-	if len(schedPod.Spec.Affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms) != 1 {
-		t.Errorf("fail to mutate pod, not add node affinity")
-	}
-}
